requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
python-jose[cryptography]>=3.3.0
pyyaml>=6.0.1
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0
cryptography>=45.0.0
httpx>=0.25.2
black>=23.0.0
ruff>=0.1.0
pytest>=7.0.0
pytest-asyncio>=0.23.0
pytest-httpx>=0.21.0